+++ src/components/ImagePanel/ImagePanel.tsx
@@ -8,9 +8,9 @@ import { css, cx } from '@emotion/css';
 import { FieldType, PanelProps } from '@grafana/data';
 import { Alert, PageToolbar, ToolbarButton, useStyles2 } from '@grafana/ui';
 import { ImageSizeModes, ImageTypesSymbols, SupportedTypes, TestIds } from '../../constants';
-import { Styles } from '../../styles';
 import { ButtonType, PanelOptions, ZoomType } from '../../types';
 import { base64toBlob } from '../../utils';
+import { Styles } from './ImagePanel.styles';
 
 /**
  * Properties
@@ -313,7 +313,14 @@ export const ImagePanel: React.FC<Props> = ({ options, data, width, height, repl
    * Add URL to Image
    */
   let image = (
-    <img width={imageWidth || ''} height={imageHeight || ''} src={img} data-testid={TestIds.panel.image} alt="" />
+    <img
+      width={imageWidth || ''}
+      height={imageHeight || ''}
+      src={img}
+      data-testid={TestIds.panel.image}
+      alt=""
+      style={{ imageRendering: options.scale }}
+    />
   );
   if (options.url) {
     const url = replaceVariables(options.url);