+++ src/constants/image.ts
@@ -1,3 +1,5 @@
+import { ImageScale } from '../types';
+
 /**
  * Image Types
  */
@@ -49,4 +51,23 @@ export const SizeModeOptions = [
   { value: ImageSizeModes.AUTO, label: 'Panel', description: 'Based on panel size' },
   { value: ImageSizeModes.ORIGINAL, label: 'Original' },
   { value: ImageSizeModes.CUSTOM, label: 'Custom' },
+];
+
+/**
+ * Image Scale Options
+ */
+export const ImageScaleOptions = [
+  { value: ImageScale.AUTO, label: 'Auto' },
+  {
+    value: ImageScale.CRISP_EDGES,
+    label: 'Crisp Edges',
+    description:
+      'The image is scaled with an algorithm that preserves contrast and edges in the image. Generally intended for images such as pixel art or line drawings, no blurring or color smoothing occurs.',
+  },
+  {
+    value: ImageScale.PIXELATED,
+    label: 'Pixelated',
+    description:
+      'The image is scaled with the "nearest neighbor" or similar algorithm, preserving a "pixelated" look as the image changes in size.',
+  },
 ];